Add files via upload
The dashboard seeks to analyse the prescription of medications to the residents in the Tees Valley region of Middlesbrough, Stockton, Redcar, Cleveland, Darlington, & Hartlepool for the year 2014
The dataset for this analysis was obtained from the NHS website below. It contains detailed information on prescriptions issued in England that have been dispensed in England, Wales, Scotland, Guernsey, Alderney, Jersey, and the Isle of Man from the year 2014 to 2022.