Add files via upload
This infographic was created using the data viz website: Vengage.

The project seeks to investigate and provide insights into how HULU compares itself in its products offerings against three other major video streaming giants in the industry.  Explore problem areas to understand the reasons why it lags behind it's competition ; and recommend new strategies to be put in place to capture new markets in the industry.