Add files via upload
This dashboard seek to analyse the progression of answers from respondents on the famous Data Science survey conducted by Kaggle from 2017 to 2021. The report seek to explore and investigate the state of data science ecosystem especially the inclusion of women in data science, the incorporation of machine learning frameworks, and the adoption of several data tools and database techniques in the industry.